Added a Graphical User Interface Renamed `types` to `particle`
Set `params.delta_time` to a fixed `1 / 60`,
which fixes an issue where lowering the framerate caused the particles to "explode".

The disadvantage being that at higher or lower framerates the simulation is faster or slower then it should be.

This should be customizable in the future, where higher framerate is more accurate but slower,
unless a better solution is found.